+++ src/middlewared/middlewared/plugins/crypto.py
@@ -6,7 +6,6 @@ import inspect
 import ipaddress
 import itertools
 import josepy as jose
-import json
 import os
 import random
 import re
@@ -19,7 +18,7 @@ import middlewared.sqlalchemy as sa
 from middlewared.validators import Email, IpAddress, Range
 from middlewared.utils import osc
 
-from acme import client, errors, messages
+from acme import errors, messages
 from OpenSSL import crypto, SSL
 from contextlib import suppress
 
@@ -1312,44 +1311,6 @@ class CertificateService(CRUDService):
         return names
 
     @private
-    def get_acme_client_and_key(self, acme_directory_uri, tos=False):
-        data = self.middleware.call_sync('acme.registration.query', [['directory', '=', acme_directory_uri]])
-        if not data:
-            data = self.middleware.call_sync(
-                'acme.registration.create',
-                {'tos': tos, 'acme_directory_uri': acme_directory_uri}
-            )
-        else:
-            data = data[0]
-        # Making key now
-        key = jose.JWKRSA.fields_from_json(json.loads(data['body']['key']))
-        key_dict = key.fields_to_partial_json()
-        # Making registration resource now
-        registration = messages.RegistrationResource.from_json({
-            'uri': data['uri'],
-            'terms_of_service': data['tos'],
-            'body': {
-                'contact': [data['body']['contact']],
-                'status': data['body']['status'],
-                'key': {
-                    'e': key_dict['e'],
-                    'kty': 'RSA',
-                    'n': key_dict['n']
-                }
-            }
-        })
-
-        return client.ClientV2(
-            messages.Directory({
-                'newAccount': data['new_account_uri'],
-                'newNonce': data['new_nonce_uri'],
-                'newOrder': data['new_order_uri'],
-                'revokeCert': data['revoke_cert_uri']
-            }),
-            client.ClientNetwork(key, account=registration)
-        ), key
-
-    @private
     def acme_issue_certificate(self, job, progress, data, csr_data):
         verrors = ValidationErrors()
 
@@ -1401,7 +1362,9 @@ class CertificateService(CRUDService):
         if verrors:
             raise verrors
 
-        acme_client, key = self.get_acme_client_and_key(data['acme_directory_uri'], data['tos'])
+        acme_client, key = self.middleware.call_sync(
+            'acme.get_acme_client_and_key', data['acme_directory_uri'], data['tos']
+        )
         try:
             # perform operations and have a cert issued
             order = acme_client.new_order(csr_data['CSR'])
@@ -2088,7 +2051,9 @@ class CertificateService(CRUDService):
         certificate = self.middleware.call_sync('certificate.get_instance', id)
 
         if certificate.get('acme'):
-            client, key = self.get_acme_client_and_key(certificate['acme']['directory'], True)
+            client, key = self.middleware.call_sync(
+                'acme.get_acme_client_and_key', certificate['acme']['directory'], True
+            )
 
             try:
                 client.revoke(
